Create longest-increasing-subsequence.py
This is an improvement of the conventional DP approach by using a dictionary to keep track of all the elements belonging to a particular score. From the nature of how the elements are appended, the list of each score has its elements in the descending order always, making it easier to search for the right score for any given element.